requirements.txt
dash
dash_bootstrap_components
Pillow
pandas
gunicorn
numpy
plotly
torchvision
torch
